reduced cognitive complexity in patient registration step one
